Moved non-compiling java file to non-compiling/checks
