+++ java-checks-test-sources/default/src/main/files/non-compiling/checks/UnusedVariablesFPCheck.java
@@ -1,4 +1,4 @@
-package com;
+package checks;
 
 
 public class UnusedVariablesFPCheck {
